AI/langchain/aws_bedrock_UC/content_creation_app/requirements.txt
boto3
pypdf
faiss-cpu
numpy
rich
